+++ components/style/lib.rs
@@ -105,6 +105,7 @@ pub mod counter_style;
 pub mod custom_properties;
 pub mod data;
 pub mod dom;
+pub mod driver;
 pub mod element_state;
 #[cfg(feature = "servo")] mod encoding_support;
 pub mod error_reporting;
@@ -128,7 +129,6 @@ pub mod sharing;
 pub mod style_resolver;
 pub mod stylist;
 #[cfg(feature = "servo")] #[allow(unsafe_code)] pub mod servo;
-pub mod sequential;
 pub mod str;
 pub mod style_adjuster;
 pub mod stylesheet_set;
